[iOS] Test: Test코드 중복함수 수정 및 폴더 구분
