+++ iOS/IssueTracker/IssueTrackerUnitTests/Extension/LoginResponse+Encodable.swift
@@ -1,0 +1,20 @@
+//
+//  LoginResponse+Encodable.swift
+//  IssueTrackerUnitTests
+//
+//  Created by 최동규 on 2020/11/05.
+//
+
+import Foundation
+@testable import IssueTracker
+
+extension LoginResponse: Encodable {
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try container.encode(token, forKey: .token)
+    }
+    
+    enum CodingKeys: String, CodingKey {
+        case token
+    }
+}
