+++ iOS/IssueTracker/IssueTrackerUnitTests/Extension/MileStone+Encodable.swift
@@ -1,0 +1,22 @@
+//
+//  MileStone+Encodable.swift
+//  IssueTrackerUnitTests
+//
+//  Created by 최동규 on 2020/11/05.
+//
+
+import Foundation
+@testable import IssueTracker
+
+extension MileStone: Encodable {
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try container.encode(id, forKey: .id)
+        try container.encode(title, forKey: .title)
+    }
+    
+    enum CodingKeys: String, CodingKey {
+        case id
+        case title
+    }
+}
