+++ iOS/IssueTracker/IssueTrackerUnitTests/Info.plist
