+++ iOS/IssueTracker/IssueTrackerUnitTests/SignUpUseCaseTests.swift
@@ -1,8 +1,8 @@
 //
 //  SignUpUseCaseTests.swift
-//  SignUpUseCaseTests
+//  IssueListUseCaseTests
 //
-//  Created by 최동규 on 2020/10/29.
+//  Created by 최동규 on 2020/11/05.
 //
 
 import XCTest
@@ -16,15 +16,6 @@ final class SignUpUseCaseTests: XCTestCase {
         func request(requestType: RequestType, completionHandler: @escaping (Result<Data, NetworkError>) -> Void) {
             completionHandler(.success(Data()))
         }
-    }
-    
-    struct MockFailureNetworkService: NetworkServiceProviding {
-        var userToken: String?
-        
-        func request(requestType: RequestType, completionHandler: @escaping (Result<Data, NetworkError>) -> Void) {
-            completionHandler(.failure(.invalidData))
-        }
-        
     }
     
     let signUpInfo = SignUpInfo(email: "test@naver.com",
